Break `BeaconState` tests into separate file
